Updated Service Worker Registration
